Move file loader code into directories.
Not making sense to put these all in the same fiel anymore.
